pyproject.toml
[tool.poetry]
name = "npc-bot"
version = "0.7.0"
description = "Discord bot powered by OpenAI."
authors = ["Pragma8123 <max@machon.family>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.343"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
openai = "^1.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
